Every page on KBTree has a split type
Now that we have a split type for the nodes of the k-b-tree, we need to
reimplement the splitting strategy, as follows:

 - when splitting a page, we'll pick the median point (on the coordinate
selected by the split type), and we'll divide the page in 2 according to
that table

NOTE: I'm thinking it might be useful to expand the pagesize parameter
into 2 parameters: pagesize for leafs and pagesize for regions. Since it
looks like me that they should be treated differently.