New splitting+insertion strategy for k-b-tree
Leaf nodes will have a maximum of `pagesize` elements, but every other
node in the tree will be limited to `pagesize / 2`. That way, whenever a
split is required, all the new leaf nodes will have at least 2 elements.

To find the child where to insert a given point, we look for the last
child that wouldn't be forced to change its lower boundary for the given
coordinate.